+++ tests/Rector/Typehint/ParentTypehintedArgumentRector/Correct/correct.php.inc
