+++ tests/Rector/Typehint/ParentTypehintedArgumentRector/Correct/correct2.php.inc
