+++ tests/Rector/Typehint/ParentTypehintedArgumentRector/ParentTypehintedArgumentRectorTest.php
@@ -1,14 +1,11 @@
 <?php declare(strict_types=1);
 
-namespace Rector\Tests\Rector\Dynamic\ReturnTypehintRector;
+namespace Rector\Tests\Rector\Typehint\ParentTypehintedArgumentRector;
 
 use Iterator;
 use Rector\Testing\PHPUnit\AbstractRectorTestCase;
 
-/**
- * @see \Rector\Rector\Dynamic\ReturnTypehintRector
- */
-final class ReturnTypehintRectorTest extends AbstractRectorTestCase
+final class ParentTypehintedArgumentRectorTest extends AbstractRectorTestCase
 {
     /**
      * @dataProvider provideWrongToFixedFiles()
@@ -22,9 +19,6 @@ final class ReturnTypehintRectorTest extends AbstractRectorTestCase
     {
         yield [__DIR__ . '/Wrong/wrong.php.inc', __DIR__ . '/Correct/correct.php.inc'];
         yield [__DIR__ . '/Wrong/wrong2.php.inc', __DIR__ . '/Correct/correct2.php.inc'];
-        yield [__DIR__ . '/Wrong/wrong3.php.inc', __DIR__ . '/Correct/correct3.php.inc'];
-        yield [__DIR__ . '/Wrong/wrong4.php.inc', __DIR__ . '/Correct/correct4.php.inc'];
-        yield [__DIR__ . '/Wrong/wrong5.php.inc', __DIR__ . '/Correct/correct5.php.inc'];
     }
 
     protected function provideConfig(): string
