+++ tests/Rector/Typehint/ParentTypehintedArgumentRector/Wrong/wrong.php.inc
