+++ tests/Rector/Typehint/ParentTypehintedArgumentRector/Wrong/wrong2.php.inc
