+++ tests/Rector/Typehint/ReturnTypehintRector/Correct/correct.php.inc
