+++ tests/Rector/Typehint/ReturnTypehintRector/Correct/correct2.php.inc
