+++ tests/Rector/Typehint/ReturnTypehintRector/Correct/correct3.php.inc
