+++ tests/Rector/Typehint/ReturnTypehintRector/Correct/correct4.php.inc
