+++ tests/Rector/Typehint/ReturnTypehintRector/Correct/correct5.php.inc
