+++ tests/Rector/Typehint/ReturnTypehintRector/ReturnTypehintRectorTest.php
@@ -1,11 +1,14 @@
 <?php declare(strict_types=1);
 
-namespace Rector\Tests\Rector\Dynamic\ParentTypehintedArgumentRector;
+namespace Rector\Tests\Rector\Typehint\ReturnTypehintRector;
 
 use Iterator;
 use Rector\Testing\PHPUnit\AbstractRectorTestCase;
 
-final class ParentTypehintedArgumentRectorTest extends AbstractRectorTestCase
+/**
+ * @see \Rector\Rector\Typehint\ReturnTypehintRector
+ */
+final class ReturnTypehintRectorTest extends AbstractRectorTestCase
 {
     /**
      * @dataProvider provideWrongToFixedFiles()
@@ -19,6 +22,9 @@ final class ParentTypehintedArgumentRectorTest extends AbstractRectorTestCase
     {
         yield [__DIR__ . '/Wrong/wrong.php.inc', __DIR__ . '/Correct/correct.php.inc'];
         yield [__DIR__ . '/Wrong/wrong2.php.inc', __DIR__ . '/Correct/correct2.php.inc'];
+        yield [__DIR__ . '/Wrong/wrong3.php.inc', __DIR__ . '/Correct/correct3.php.inc'];
+        yield [__DIR__ . '/Wrong/wrong4.php.inc', __DIR__ . '/Correct/correct4.php.inc'];
+        yield [__DIR__ . '/Wrong/wrong5.php.inc', __DIR__ . '/Correct/correct5.php.inc'];
     }
 
     protected function provideConfig(): string
