+++ tests/Rector/Typehint/ReturnTypehintRector/Wrong/wrong.php.inc
