+++ tests/Rector/Typehint/ReturnTypehintRector/Wrong/wrong2.php.inc
