+++ tests/Rector/Typehint/ReturnTypehintRector/Wrong/wrong3.php.inc
