+++ tests/Rector/Typehint/ReturnTypehintRector/Wrong/wrong4.php.inc
