+++ tests/Rector/Typehint/ReturnTypehintRector/Wrong/wrong5.php.inc
